Handle text input
